+++ frontend/src/components/dashboardcomponents/BigNumber.tsx
