+++ frontend/src/components/dashboardcomponents/GraphTypeButtons.tsx
