+++ frontend/src/components/dashboardcomponents/Slider.tsx
